requirements.txt
django>=4.2,<=4.3
djangorestframework==3.14.0
Markdown==3.5.1
django-filter==23.4